Finale code reorganisation in App.js
