+++ Project-I-Social-Card/social-card/src/components/CardComponents/CardBanner.js
@@ -1,2 +1,10 @@
 import React from 'react';
 import './Card.css';
+
+function CardImage ({cardImgLink}) {
+    return (
+      <img src={cardImgLink} alt=""/>
+    );
+  }
+
+  export default CardImage;
