+++ Project-I-Social-Card/social-card/src/components/CardComponents/CardContent.js
@@ -1,2 +1,18 @@
 import React from 'react';
 import './Card.css';
+
+function CardContent () {
+    return (
+      <div className="cardContent">
+        <h2>
+          Get started with react
+        </h2>        
+        <p>
+          React make painless to create interactives UIs. Design simple view for each state in your application.
+        </p>
+        <span>reactjs.org</span>
+      </div>
+    );
+  }
+
+  export default CardContent;
